Session, OTP and Emails working for login

+++ Admin/delete_rescue.php
@@ -1,4 +1,7 @@
 <?php	include("database.php");
+if(!isset($_SESSION['username']) || !isset($_SESSION['otp'])) {
+          header('location:Adminlogin.php');
+        }
 		$rid = $_POST['rescue_id'];
 		foreach($rid as $val) {
 			$sql = "delete from rescue where rid =".$val;
